Updated users table and password reset route.

+++ data/migrations/20201105140441_add-tables.js
@@ -21,7 +21,7 @@ exports.up = function (knex) {
       tbl.string("website_url", 255);
       tbl.string("header_img", 255);
       tbl.string("reset_password_token", 255);
-      tbl.integer("reset_password_expires");
+      tbl.bigInteger("reset_password_expires");
     })
     .createTable("subscribers", (tbl) => {
       tbl.increments();
